+++ peepholelib/datasets/imagenet.py
@@ -88,19 +88,3 @@ class ImageNet(DatasetWrap):
             meta = pickle.load(f, encoding='latin1')
         labels = {i: name for i, name in enumerate(meta['fine_label_names'])}
         return labels 
-    
-    def get(self, ds_key, idx):
-        '''
-        Get item from the dataset.
-        
-        Args:
-        - idx (int): Index of the item to get.
-        - ds_key (str): Key of the dataset to get the item from ('train', 'val', 'test').
-        
-        Returns:
-        - a tuple of (image, label)
-        '''
-        if not self.__dataset__:
-            raise RuntimeError('Data not loaded. Please run load_data() first.')
-        
-        return [self.__dataset__[ds_key][idx]]